x/gentlemint/types/message_buy_shr_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/sharering/shareledger/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgBuyShr_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgBuyShr
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgBuyShr{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgBuyShr{
				Creator: sample.AccAddress(),
				Amount:  "1",
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
